refactor: 過去に定義した Dao や Entity を legacy パッケージに移した
