+++ core/database/src/main/java/app/doggy/core/database/legacy/entity/BookAuthorCrossRef.kt
@@ -1,4 +1,4 @@
-package app.doggy.core.database.entity
+package app.doggy.core.database.legacy.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
